Bump version, and because I don't like bigger than 0.9 versions, then:
This is 1.0!!!

This project is the best because of you guys that added support to it
for many many things, I wanna thank everyone who download it, I know
it could be just one bash command, but making things easier with
other software is the way UNIX told us to do, and this project
is the best example of it.

Thank you guys!